+++ CHANGELOG.md
@@ -6,7 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next 2022-10-14
+
+### Fixed
+
+On DALnet, fixed auto-restart after net-split was not properly detecting the nick collision.
+
 ## [v0.2.20](https://github.com/cotarr/irc-hybrid-client/releases/tag/v0.2.20) 2022-09-23
+
+### Changed
+
+- Dev dependencies, updated eslint@8.25.0 including eslint dependencies
 
 ### Changed